adf-aiml-pipeline/requirements.txt
azure-identity==1.15.0
# Flask[async]==2.3.2
openai==1.6.1
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0
quart==0.19.4
uvicorn==0.24.0
aiohttp==3.9.2
gunicorn==20.1.0
azure-ai-formrecognizer==3.2.1
Markdown==3.4.4
requests==2.31.0
tqdm==4.66.1
tiktoken==0.4.0
langchain==0.0.340
bs4==0.0.1
urllib3==2.1.0
pytest==7.4.0
pytest-asyncio==0.23.2
azure-storage-blob
chardet
azure-keyvault